requirements.txt
# Core dependencies
click>=8.0.0
requests>=2.25.0
pydantic>=2.0.0
python-dotenv>=0.19.0
rich>=13.0.0
keyring>=23.0.0
cryptography>=3.4.0
pyjwt>=2.4.0
pytz>=2023.3
pyyaml>=6.0.0

# Web UI dependencies
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.8.0

# Database dependencies
sqlalchemy>=2.0.0
alembic>=1.12.0

# Dataset dependencies
datasets>=2.14.0

# Optional dependencies for development
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
isort>=5.10.0
flake8>=5.0.0
mypy>=0.991
pre-commit>=2.20.0
pytest-mock>=3.10.0
responses>=0.22.0
